+++ lib/cpp/mosquittopp.cpp
@@ -38,7 +38,7 @@ static void on_connect_wrapper(void *obj, int rc)
 	m->on_connect(rc);
 }
 
-static void on_publish_wrapper(void *obj, int mid)
+static void on_publish_wrapper(void *obj, uint16_t mid)
 {
 	class mosquittopp *m = (class mosquittopp *)obj;
 	m->on_publish(mid);
@@ -50,13 +50,13 @@ static void on_message_wrapper(void *obj, struct mosquitto_message *message)
 	m->on_message(message);
 }
 
-static void on_subscribe_wrapper(void *obj, int mid, int qos_count, uint8_t *granted_qos)
+static void on_subscribe_wrapper(void *obj, uint16_t mid, int qos_count, uint8_t *granted_qos)
 {
 	class mosquittopp *m = (class mosquittopp *)obj;
 	m->on_subscribe(mid, qos_count, granted_qos);
 }
 
-static void on_unsubscribe_wrapper(void *obj, int mid)
+static void on_unsubscribe_wrapper(void *obj, uint16_t mid)
 {
 	class mosquittopp *m = (class mosquittopp *)obj;
 	m->on_unsubscribe(mid);
